--- OpenHD/ohd_discovery/.gitignore
@@ -1,15 +1,0 @@
-CMakeLists.txt.user
-CMakeCache.txt
-CMakeFiles
-CMakeScripts
-Testing
-x_Makefile
-cmake_install.cmake
-install_manifest.txt
-compile_commands.json
-CTestTestfile.cmake
-_deps
-
-.idea
-cmake-build-createDebug
-build
